issue: 4082814 Adding DOCA send failure handling for future TLS
It is crucial to revert expected seq num in case of TX drop for TLS offload.
Adding this handling now to avoid serious debugging in the future.

+++ src/core/sock/sockinfo_nvme.h
@@ -66,7 +66,7 @@ public:
     int setsockopt(int __level, int __optname, const void *__optval, socklen_t __optlen) override;
     ssize_t tx(xlio_tx_call_attr_t &tx_arg) override;
     int postrouting(struct pbuf *p, struct tcp_seg *seg, xlio_send_attr &attr) override;
-    bool handle_send_ret(ssize_t ret, struct tcp_seg *seg) override;
+    bool handle_send_ret(uint32_t ret, struct tcp_seg *seg) override;
     err_t recv(struct pbuf *p) override;
 
 private:
